# Spring Boot Assessment Swagger v1.0.0.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Spring Boot Assessment APIs"
paths:
  /products:
    get:
      summary: "Get all products"
      responses:
        200:
          description: "Successful operation"
    post:
      summary: "Create a new product"
      parameters:
        - name: "product"
          in: "body"
          description: "Product object"
          required: true
          schema:
            $ref: "#/definitions/Product"
      responses:
        201:
          description: "Product created successfully"
  /products/{productId}:
    get:
      summary: "Get a product by ID"
      parameters:
        - name: "productId"
          in: "path"
          description: "ID of the product to retrieve"
          required: true
          type: "integer"
      responses:
        200:
          description: "Successful operation"
        404:
          description: "Product not found"
    put:
      summary: "Update a product by ID"
      parameters:
        - name: "productId"
          in: "path"
          description: "ID of the product to update"
          required: true
          type: "integer"
        - name: "product"
          in: "body"
          description: "Updated product object"
          required: true
          schema:
            $ref: "#/definitions/Product"
      responses:
        200:
          description: "Product updated successfully"
        404:
          description: "Product not found"
    delete:
      summary: "Delete a product by ID"
      parameters:
        - name: "productId"
          in: "path"
          description: "ID of the product to delete"
          required: true
          type: "integer"
      responses:
        204:
          description: "Product deleted successfully"
        404:
          description: "Product not found"
definitions:
  Product:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      description:
        type: "string"